basic division (needs to be improved)
